requirements.txt
coverage==4.5.4
docker==4.0.2
flake8==3.5.0
isort==4.3.4
-e .[sa]
sphinx==2.2.0
tox==3.14.0
pytest==5.1.3
pytest-cov==2.7.1
pytest-sugar==0.9.2
pytest-timeout==1.3.3
sphinxcontrib-asyncio==0.2.0
psycopg2-binary==2.7.7
sqlalchemy[postgresql_psycopg2binary]==1.3.8
